codec: bench: streamline benchmarking
bench_test is the "helper" file for all benchmarking tests.

There are also benchmarks which depend on just codec and the stdlib,
and benchmarks which depend on external libraries.
It is an explicit goal that you can run benchmarks without external
dependencies (which is why the 'x' build tag was explicitly introduced).

There are 2 ways of running tests:
   - generated
   - not generated

Consequently, we have 4 groups:
   - codec_bench   (gen, !gen)
   - stdlib_bench  (!gen only)
   - x_bench       (!gen only)
   - x_bench_gen   (gen only)

We also have 4 matching suite files.
   - z_all_bench (rename later to z_all_codec_bench???)
   - z_all_stdlib_bench
   - z_all_x_bench
   - z_all_x_bench_gen

Finally, we have a single test (TestBenchInit) that
will log information about whether each format can
encode or not, how long to encode (unscientifically),
and the encode size.

This test MUST be run always, as it calls init() internally